phase8-s1/s5: fused megakernel — no-hang barrier + real bwd->opt SMEM fusion
- barrier hang fixed (occupancy-cap + co-residency assert): launch_fused_megakernel
  queries cudaOccupancyMaxActiveBlocksPerMultiprocessor, pins 1 CTA/SM, and returns
  cudaErrorLaunchOutOfResources (NOT hang) if occ==0; counters device-zeroed each
  launch. gfx942 twin via hipOccupancy*. (Chosen over cooperative launch to keep the
  reusable hand-built GridBarrier.)
- int64 indexing in model_stages/opt_components/fused_megakernel offsets (off+i,
  rebase_state, float4 tail) — apply_optimizer idx int64 (algorithms already int64)
- TRUE cross-phase fusion (realized): bwd grad computed into __shared__ smem_grad
  tile and consumed in-place by apply_optimizer -> grad never round-trips global,
  bwd->opt grid barrier ELIMINATED (L3: 2 barriers -> 1). Element-local + one CTA
  owns a tensor end-to-end (work-stealing safe). Bit-identical grad. 🟡 perf.
  DEFERRED (documented): fwd 'acts' stays global (fwd->bwd barrier + queue reset
  can reassign the owning CTA; SMEM acts can't cross safely without abandoning
  work-stealing / DSMEM). psi_b2 out of scope (codegen-template item).

Verified: 4 L3 cells (decoder/adamw, mamba3/muon, decoder/neuralgrok, decoder/sg2)
COMPILE_OK; gfx942 fused_megakernel.hip.hpp AMDGCN_OK; drift guard PASS.

+++ csrc/fused/gfx942/opt_components.hip.hpp
@@ -77,7 +77,7 @@ struct FusedOptState {
 // Shared Adam tail: m,v EMAs on g_eff, bias-corrected decoupled-WD apply.
 // (bc1/bc2 un-inverted = 1-beta^t; divide — matches the algorithm headers.)
 __device__ __forceinline__ float sg_adam_tail(
-        float p, float g_eff, float* m_buf, float* v_buf, int i,
+        float p, float g_eff, float* m_buf, float* v_buf, long long i,
         float b1, float b2, float eps, float bc1, float bc2,
         float lr, float wd) {
     const float m = b1 * m_buf[i] + (1.0f - b1) * g_eff;
@@ -112,7 +112,7 @@ __device__ __forceinline__ float sg_sg15_alpha(float mu_val, float ab, float am)
 template <OptId Opt>
 __device__ __forceinline__ void apply_optimizer(
         float* __restrict__ params, const float* __restrict__ grad,
-        int i, int step, const FusedOptState& st) {
+        long long i, int step, const FusedOptState& st) {
     (void)step;
     const float g = grad[i];
     const float p = params[i];
